Stage 5: restore float64 for maxind
One line. maxind was np.float_ -- which IS float64 -- until commit 7c19787 ("Updated
numpy datatypes") swapped it for np.float32 while removing aliases numpy 2 dropped.
That was a silent precision regression shipped as an alias fix, not a decision, so
this is a revert rather than a change.

Golden diff: ZERO data voxels. That is expected, not luck, and the reason the
regression survived two years unnoticed:

  * step_size_inv is ALREADY float64 -- it is computed as
    1.0 * (number_of_pixels - 1) / maxind / 2 with number_of_pixels an int64 array,
    and numpy promotes int64/float32 to float64. So the cast never reached the
    reciprocal step, only the additive offset and the stored limit.
  * maxind values that float32 represents exactly (7.0, 5.0, 1.5 -- i.e. almost
    everything anyone types) were bit-identical either way. Measured over 500k random
    coordinates on an 801^3 grid: 0 voxels differ. Only unrepresentable values like
    7.3 or 6.7 moved anything, and then ~0.001% of coordinates.

What it does buy: lower_limits = -maxind is written to the output and Yell reads it
to place the grid in reciprocal space. Under float32, maxind=7.3 was stored as
-7.3000001907 -- off by 1.9e-7 r.l.u. Now exact. The golden's lower_limits dtype goes
float32 -> float64; its values do not change.

Corrects the plan, which claimed float32's "~7 digits are spent right where the index
is computed". They are not. Stated here so the claim is not re-inflated later.

The two Stage-1 marker tests invert as designed. rebinned_data stays float32
deliberately -- at 801^3 that is 2.06 GB against 4.11 GB, which matters on a 16 GB
machine -- so the guard names maxind specifically rather than banning float32.

+++ meerkat/meerkat.py
@@ -352,7 +352,11 @@ def reconstruct_data(filename_template,
     number_of_pixels = np.array(number_of_pixels)
     assert len(number_of_pixels) == 3
 
-    maxind = np.array(maxind, dtype=np.float32)
+    # float64, not float32. This was np.float_ (i.e. float64) until commit 7c19787
+    # swapped it for float32 while removing numpy-2 aliases -- a silent precision
+    # regression, not a decision. maxind is written straight to the output as
+    # lower_limits = -maxind, which Yell reads to place the grid in reciprocal space.
+    maxind = np.array(maxind, dtype=np.float64)
     assert len(maxind) == 3
 
     step_size_inv = 1.0 * (number_of_pixels - 1) / maxind / 2
